+++ R/query-set-op.R
@@ -51,3 +51,46 @@ sql_render.set_op_query <- function(query,
     # nocov end
   }
 }
+
+#' @export
+#' @rdname sql_build
+union_query <- function(x, unions) {
+  structure(
+    list(
+      x = x,
+      unions = unions
+    ),
+    class = c("union_query", "query")
+  )
+}
+
+#' @export
+print.union_query <- function(x, ...) {
+  cat_line(indent_print(sql_build(x$x)))
+
+  for (i in seq_along(x$unions$table)) {
+    if (x$unions$all[[i]]) {
+      method <- "UNION ALL"
+    } else {
+      method <- "UNION"
+    }
+    cat_line()
+    cat_line(indent(sql(method)))
+    cat_line()
+
+    cat_line(indent_print(sql_build(x$unions$table[[i]])))
+  }
+}
+
+#' @export
+sql_render.union_query <- function(query, con = NULL, ..., subquery = FALSE, lvl = 0) {
+  from_x <- sql_render(query$x, con, ..., subquery = FALSE, lvl = lvl)
+  unions <- list()
+  unions$table <- purrr::map(
+    query$unions$table,
+    function(table) sql_render(table, con, ..., subquery = FALSE, lvl = lvl)
+  )
+  unions$all <- query$unions$all
+
+  sql_query_union(con, from_x, unions, lvl = lvl)
+}