+++ tests/testthat/test-verb-mutate.R
@@ -440,6 +440,12 @@ test_that("mutate() uses star", {
     lf %>% transmute(a = 1L, x, y, z = 2L) %>% remote_query(),
     sql("SELECT 1 AS `a`, *, 2 AS `z`\nFROM `df`")
   )
+
+  # does not use * if `use_star = FALSE`
+  expect_equal(
+    lf %>% mutate(z = 1L) %>% remote_query(use_star = FALSE),
+    sql("SELECT `x`, `y`, 1 AS `z`\nFROM `df`")
+  )
 })
 
 # sql_build ---------------------------------------------------------------